feat: move app data state to custom data state hook
